Extract support generators for modular code generation (Issue #185)
ADR-053 A5: Extract support/utility code from CodeGenerator.ts into
modular generator files following the strangler fig pattern.

New files created:
- src/codegen/types/TYPE_MAP.ts - C-Next to C type mapping
- src/codegen/types/WIDER_TYPE_MAP.ts - Wider types for safe arithmetic
- src/codegen/types/TYPE_LIMITS.ts - TYPE_MAX and TYPE_MIN constants
- src/codegen/generators/support/HelperGenerator.ts - Overflow & safe-div helpers
- src/codegen/generators/support/IncludeGenerator.ts - Include/directive processing
- src/codegen/generators/support/CommentUtils.ts - Comment handling utilities
- src/codegen/generators/support/index.ts - Barrel export

CodeGenerator.ts now delegates to these extracted modules while
maintaining the same public interface.
